Accept files under 1GB
As we discussed on Slack, we'll be more welcoming to visual novels, which typically run in the range of hundreds of megabytes. Previously we said that we'd ask questions about files over 100MB, (i.e. most VNs).

This PR raises the "no questions asked" limit to 1GB, which should cover the vast majority of VNs.